Add support for message/rfc822 body parts
The existing Text and Binary body part types always ends up with a
Content-Transfer-Encoding of base64, quoted-printable, or, in certain
circumstances, 7bit. None of these are suitable for attaching a full
e-mail message. RFC2046 states: "No encoding other than "7bit", "8bit",
or "binary" is permitted for the body of a "message/rfc822" entity".

This patch adds support for a BodyPart::Message type, which will always
be rendered with a Content-Transfer-Encoding of "binary".